fix telegram hook
